+++ db/migrate/20150324193241_add_position_to_schedules.rb
@@ -1,0 +1,5 @@
+class AddPositionToSchedules < ActiveRecord::Migration
+  def change
+    add_column :schedules, :position, :integer, null: false, default: 0
+  end
+end