+++ test/controllers/subjects_controller_test.rb
@@ -7,6 +7,11 @@ class SubjectsControllerTest < ActionController::TestCase
     @subject = subjects(:one)
   end
 
+  test "should get report" do
+    get :report, project_id: @project
+    assert_response :success
+  end
+
   test "should get index" do
     get :index, project_id: @project
     assert_response :success